Reworked title and front rendering
The title, image and subtitle are now Flowables that go into frames. To allow this the front of the card was redesigned to use a frame for layout.

Sadly a hack was required to work around the fact the card need to be drawn bottom to top but frames only support top to bottom.